Update server app to fetch NCES.json file from S3 bucket before starting Express app, and caching the data on the Express app's locals object
