+++ app/server/app/routes/formioNCES.js
@@ -1,12 +1,8 @@
-const { resolve } = require("node:path");
-const { readFile } = require("node:fs/promises");
 const express = require("express");
-const axios = require("axios").default || require("axios"); // TODO: https://github.com/axios/axios/issues/5011
 // ---
-const { s3BucketUrl } = require("../config/s3");
 const log = require("../utilities/logger");
 
-const { NODE_ENV, FORMIO_NCES_API_KEY } = process.env;
+const { FORMIO_NCES_API_KEY } = process.env;
 
 const router = express.Router();
 
@@ -42,41 +38,16 @@ router.get("/:searchText?", (req, res) => {
     return res.json({});
   }
 
-  const localFilePath = resolve(__dirname, "../content", "nces.json");
-  const s3FileUrl = `${s3BucketUrl}/content/nces.json`;
-  const logMessage = `Fetching NCES.json from S3 bucket.`;
+  const result = req.app.locals.ncesData.find((item) => {
+    return item["NCES ID"] === searchText;
+  });
 
-  Promise.resolve(
-    /**
-     * local development: read file directly from disk
-     * Cloud.gov: fetch file from the public s3 bucket
-     */
-    NODE_ENV === "development"
-      ? readFile(localFilePath, "utf8").then((string) => JSON.parse(string))
-      : (log({ level: "info", message: logMessage, req }),
-        axios.get(s3FileUrl).then((res) => res.data)),
-  )
-    .then((data) => {
-      const result = data.find((item) => item["NCES ID"] === searchText);
+  const logMessage =
+    `NCES data searched with NCES ID '${searchText}' resulting in ` +
+    `${result ? "a match" : "no matches"}.`;
+  log({ level: "info", message: logMessage, req });
 
-      const logMessage =
-        `NCES data searched with NCES ID '${searchText}' resulting in ` +
-        `${result ? "a match" : "no matches"}.`;
-      log({ level: "info", message: logMessage, req });
-
-      return res.json({ ...result });
-    })
-    .catch((error) => {
-      const errorStatus = error.response?.status || 500;
-      const errorMethod = error.response?.config?.method?.toUpperCase();
-      const errorUrl = error.response?.config?.url;
-
-      const logMessage = `S3 Error: ${errorStatus} ${errorMethod} ${errorUrl}`;
-      log({ level: "error", message: logMessage, req });
-
-      const errorMessage = `Error getting NCES.json data from S3 bucket.`;
-      return res.status(errorStatus).json({ message: errorMessage });
-    });
+  return res.json({ ...result });
 });
 
 module.exports = router;
